requirements.txt
python-telegram-bot>=21.0
flask>=3.0
flask-cors>=4.0
waitress>=3.0
